refactor: move reviews & top case in packages
