+++ presentation/src/main/java/com/madrid/presentation/screens/detailsScreen/castDetails/ActorDetails.kt
@@ -36,7 +36,7 @@ import com.madrid.presentation.component.movioCards.MovioVerticalCard
 import com.madrid.presentation.navigation.Destinations
 import com.madrid.presentation.navigation.LocalNavController
 import com.madrid.presentation.viewModel.detailsViewModel.actor.ActorDetailsViewModel
-import com.madrid.presentation.viewModel.detailsViewModel.movie.MovieDetailsUiState
+import com.madrid.presentation.viewModel.detailsViewModel.actor.NetworkDetailsUiState
 
 
 @Composable
@@ -91,7 +91,7 @@ fun ActorDetails(
 
 @Composable
 private fun ActorDetailsContent(
-    actor: MovieDetailsUiState.CastUiState,
+    actor: NetworkDetailsUiState.CastUiState,
     onBackClick: () -> Unit,
     onKnownForClick: (Int) -> Unit
 ) {
