+++ presentation/src/main/java/com/madrid/presentation/screens/detailsScreen/reviewsScreen/ReviewScreen.kt
@@ -22,8 +22,8 @@ import com.madrid.designSystem.component.TopAppBar
 import com.madrid.presentation.R
 import com.madrid.presentation.navigation.LocalNavController
 import com.madrid.presentation.screens.detailsScreen.reviewsScreen.composables.ReviewCard
-import com.madrid.presentation.viewModel.detailsViewModel.ReviewsScreenUiState
-import com.madrid.presentation.viewModel.detailsViewModel.ReviewsScreenViewModel
+import com.madrid.presentation.viewModel.detailsViewModel.review.ReviewsScreenUiState
+import com.madrid.presentation.viewModel.detailsViewModel.review.ReviewsScreenViewModel
 
 @Composable
 fun ReviewsScreen(viewModel: ReviewsScreenViewModel = hiltViewModel()) {
