+++ presentation/src/main/java/com/madrid/presentation/screens/detailsScreen/reviewsScreen/composables/ReviewsSection.kt
@@ -15,8 +15,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.madrid.designSystem.component.CustomTextTitle
 import com.madrid.presentation.R
-import com.madrid.presentation.viewModel.detailsViewModel.ReviewUiState
-import com.madrid.presentation.viewModel.detailsViewModel.ReviewsScreenUiState
+import com.madrid.presentation.viewModel.detailsViewModel.review.ReviewUiState
+import com.madrid.presentation.viewModel.detailsViewModel.review.ReviewsScreenUiState
 
 @Composable
 fun ReviewScreen(
