+++ presentation/src/main/java/com/madrid/presentation/screens/detailsScreen/seriesDetails/SeriesDetailsScreen.kt
@@ -72,8 +72,8 @@ import com.madrid.presentation.screens.detailsScreen.castDetails.LoadingScreen
 import com.madrid.presentation.screens.detailsScreen.reviewsScreen.composables.ReviewScreen
 import com.madrid.presentation.screens.detailsScreen.similarMedia.SimilarSeries
 import com.madrid.presentation.screens.detailsScreen.similarMedia.SimilarSeriesSection
-import com.madrid.presentation.viewModel.detailsViewModel.ReviewUiState
-import com.madrid.presentation.viewModel.detailsViewModel.ReviewsScreenUiState
+import com.madrid.presentation.viewModel.detailsViewModel.review.ReviewUiState
+import com.madrid.presentation.viewModel.detailsViewModel.review.ReviewsScreenUiState
 import com.madrid.presentation.viewModel.detailsViewModel.SeriesDetailsViewModel
 import com.madrid.presentation.viewModel.shared.parser.formatFullDateKtx
 import com.madrid.presentation.viewModel.shared.parser.formatYearKtx
