+++ presentation/src/main/java/com/madrid/presentation/viewModel/detailsViewModel/SeriesDetailsUiState.kt
@@ -1,5 +1,7 @@
 package com.madrid.presentation.viewModel.detailsViewModel
 
+import com.madrid.presentation.viewModel.detailsViewModel.review.ReviewUiState
+
 data class SeriesDetailsUiState(
     val seriesId: Int = 0,
     val isFavourite: Boolean = false,
