+++ presentation/src/main/java/com/madrid/presentation/viewModel/detailsViewModel/review/ReviewsScreenViewModel.kt
@@ -1,4 +1,4 @@
-package com.madrid.presentation.viewModel.detailsViewModel
+package com.madrid.presentation.viewModel.detailsViewModel.review
 
 import android.util.Log
 import androidx.lifecycle.SavedStateHandle
@@ -7,6 +7,7 @@ import com.madrid.domain.usecase.movie.GetMovieReviewsUseCase
 import com.madrid.domain.usecase.series.GetSeriesReviewsUseCase
 import com.madrid.presentation.navigation.Destinations
 import com.madrid.presentation.viewModel.base.BaseViewModel
+import com.madrid.presentation.viewModel.detailsViewModel.toUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
